requirements.txt
flask
python-qbittorrent
python-dotenv
flask-session
bcrypt
flask-bcrypt
requests
webauthn
pydantic
cryptography
PyJWT